requirements.txt
# Core ML / Audio 
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.1
soundfile>=0.12.1

# Speech Models 
openai-whisper>=20231117
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0

# G2P / Phonetics 
phonemizer>=3.2.1
epitran>=1.24
gruut>=2.3.4

# Translation / NLP 
sentencepiece>=0.1.99
sacrebleu>=2.3.1
nltk>=3.8.1

# LID / Embeddings 
speechbrain>=0.5.16
pyannote.audio>=3.1.1

# TTS (VITS) 
TTS>=0.22.0

# Anti-Spoofing / Features 
spafe>=0.3.2
sklearn-compat>=0.1.1
scikit-learn>=1.3.0

# Signal Processing 
noisereduce>=3.0.0
pydub>=0.25.1
webrtcvad>=2.0.10

# Video / Data Scraping 
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0

# DTW 
dtaidistance>=2.3.10
fastdtw>=0.3.4

# Utilities 
tqdm>=4.66.0
pyyaml>=6.0.1
omegaconf>=2.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
